conditions for animator nearly ok, missing the trigger condition

+++ FreeEngine/CameraManager.cpp
@@ -14,7 +14,7 @@ CameraManager::~CameraManager()
 {
 }
 
-void CameraManager::Move(int x, int y)
+void CameraManager::Move(float x, float y)
 {
 	_mainView.move(x, y);
 }
@@ -26,13 +26,13 @@ sf::View CameraManager::GetView()
 
 void CameraManager::Init()
 {
-	int x = Window::Instance().GetScreenResolutionX();
-	int y = Window::Instance().GetScreenResolutionY();
+	float x = Window::Instance().GetScreenResolutionX();
+	float y = Window::Instance().GetScreenResolutionY();
 	_mainView.reset(sf::FloatRect(0, 0, x, y));
 	_currentView = _mainView;
 }
 
-void CameraManager::ResizeCurrentView(int x, int y)
+void CameraManager::ResizeCurrentView(float x, float y)
 {
 	_currentView.reset(sf::FloatRect(0, 0, x, y));
 }
